contacts 1.1

+++ wa-apps/contacts/lib/actions/groups/contactsGroupsContactSave.controller.php
@@ -5,7 +5,8 @@ class contactsGroupsContactSaveController extends waJsonController
 {
     public function execute()
     {
-        if (! ( $id = (int)waRequest::get('id'))) {
+        $ids = waRequest::request('id', array(), 'array_int');
+        if (!$ids) {
             throw new waException('Contact id not specified.');
         }
 
@@ -15,12 +16,30 @@ class contactsGroupsContactSaveController extends waJsonController
         }
 
         $groups = waRequest::post('groups', array(), 'array_int');
+        
+        $counters = array();
         $ugm = new waUserGroupsModel();
-        $ugm->delete($id, array());
-        if ($groups) {
-            $ugm->add(array_map(wa_lambda('$gid', 'return array('.$id.', $gid);'), $groups));
+        if ($this->getRequest()->request('set')) {
+            foreach ($ids as $id) {
+                $ugm->delete($id, array());
+            }
         }
-        $this->response = 'ok';
+        foreach ($ids as $id) {
+            if ($groups) {
+                $ugm->add(array_map(wa_lambda('$gid', 'return array('.$id.', $gid);'), $groups));
+            }
+        }
+        $gm = new waGroupModel();
+        foreach ($groups as $gid) {
+            $cnt = $ugm->countByField(array('group_id' => $gid));
+            $gm->updateCount($gid, $cnt);
+            $counters[$gid] = $cnt; 
+        }
+        $this->response['counters'] = $counters;
+        
+        $this->response['message'] = _w("%d user has been added", "%d users have been added", count($ids));
+        $this->response['message'] .= ' ';
+        $this->response['message'] .= _w("to %d group", "to %d groups", count($groups));
     }
 }
